+++ sdk/structs/property.go
@@ -11,6 +11,8 @@ type Property struct {
 	Name     string
 	Desc     string
 	Lte      bool
+	Read     bool
+	Write    bool
 	Schema   string
 	Type     ultipa.PropertyType
 	SubTypes []ultipa.PropertyType